+++ tests/Common/CommonCommandTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Tests\Common;
 
+use PHPUnit\Framework\Attributes\DataProviderExternal;
 use ReflectionException;
 use Throwable;
 use Yiisoft\Db\Constant\DataType;
@@ -28,11 +29,15 @@ use Yiisoft\Db\Query\Query;
 use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Schema\Column\ColumnBuilder;
 use Yiisoft\Db\Tests\AbstractCommandTest;
+use Yiisoft\Db\Tests\Provider\CommandProvider;
 use Yiisoft\Db\Tests\Support\Assert;
+use Yiisoft\Db\Tests\Support\DbHelper;
 use Yiisoft\Db\Transaction\TransactionInterface;
 
+use function array_filter;
 use function is_string;
 use function setlocale;
+use function str_starts_with;
 
 abstract class CommonCommandTest extends AbstractCommandTest
 {
@@ -486,45 +491,36 @@ abstract class CommonCommandTest extends AbstractCommandTest
         $db->close();
     }
 
-    /**
-     * @dataProvider \Yiisoft\Db\Tests\Provider\CommandProvider::createIndex
-     *
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws Throwable
-     */
-    public function testCreateIndex(
-        string $name,
-        string $tableName,
-        array|string $column,
-        string|null $indexType,
-        string|null $indexMethod,
-    ): void {
+    #[DataProviderExternal(CommandProvider::class, 'createIndex')]
+    public function testCreateIndex(array $columns, array $indexColumns, string|null $indexType, string|null $indexMethod): void
+    {
         $db = $this->getConnection();
 
         $command = $db->createCommand();
         $schema = $db->getSchema();
 
+        $tableName = 'test_create_index';
+        $indexName = 'test_index_name';
+
         if ($schema->getTableSchema($tableName) !== null) {
             $command->dropTable($tableName)->execute();
         }
 
-        $command->createTable($tableName, ['int1' => 'integer not null', 'int2' => 'integer not null'])->execute();
+        $command->createTable($tableName, $columns)->execute();
 
-        $this->assertEmpty($schema->getTableIndexes($tableName, true));
+        $count = count($schema->getTableIndexes($tableName));
+        $command->createIndex($tableName, $indexName, $indexColumns, $indexType, $indexMethod)->execute();
 
-        $command->createIndex($tableName, $name, $column, $indexType, $indexMethod)->execute();
+        $this->assertCount($count + 1, $schema->getTableIndexes($tableName));
 
-        if (is_string($column)) {
-            $column = [$column];
-        }
+        $index = array_filter($schema->getTableIndexes($tableName), static fn ($index) => !$index->isPrimary())[0];
 
-        $this->assertSame($column, $schema->getTableIndexes($tableName, true)[0]->getColumnNames());
+        $this->assertSame($indexColumns, $index->getColumnNames());
 
-        if ($indexType === 'UNIQUE') {
-            $this->assertTrue($schema->getTableIndexes($tableName, true)[0]->isUnique());
+        if ($indexType !== null && str_starts_with($indexType, 'UNIQUE')) {
+            $this->assertTrue($index->isUnique());
         } else {
-            $this->assertFalse($schema->getTableIndexes($tableName, true)[0]->isUnique());
+            $this->assertFalse($index->isUnique());
         }
 
         $db->close();
@@ -918,11 +914,6 @@ abstract class CommonCommandTest extends AbstractCommandTest
         $db->close();
     }
 
-    /**
-     * @throws Exception
-     * @throws InvalidConfigException
-     * @throws Throwable
-     */
     public function testDropTable(): void
     {
         $db = $this->getConnection();
@@ -941,6 +932,78 @@ abstract class CommonCommandTest extends AbstractCommandTest
         $command->dropTable('{{testDropTable}}')->execute();
 
         $this->assertNull($schema->getTableSchema('{{testDropTable}}', true));
+
+        $db->close();
+    }
+
+    public function testDropTableIfExistsWithExistTable(): void
+    {
+        $db = $this->getConnection();
+        $command = $db->createCommand();
+        $schema = $db->getSchema();
+
+        if ($schema->getTableSchema('{{testDropTable}}') !== null) {
+            $command->dropTable('{{testDropTable}}')->execute();
+        }
+
+        $command->createTable('{{testDropTable}}', ['id' => PseudoType::PK, 'foo' => 'integer'])->execute();
+        $this->assertNotNull($schema->getTableSchema('{{testDropTable}}', true));
+
+        $command->dropTable('{{testDropTable}}', ifExists: true)->execute();
+        $this->assertNull($schema->getTableSchema('{{testDropTable}}', true));
+
+        $db->close();
+    }
+
+    public function testDropTableIfExistsWithNonExistTable(): void
+    {
+        $db = $this->getConnection();
+        $command = $db->createCommand();
+        $schema = $db->getSchema();
+
+        if ($schema->getTableSchema('{{testDropTable}}') !== null) {
+            $command->dropTable('{{testDropTable}}')->execute();
+        }
+
+        $command->dropTable('{{testDropTable}}', ifExists: true)->execute();
+        $this->assertNull($schema->getTableSchema('{{testDropTable}}', true));
+
+        $db->close();
+    }
+
+    public function testDropTableCascade(): void
+    {
+        $db = $this->getConnection();
+        $command = $db->createCommand();
+        $schema = $db->getSchema();
+
+        if ($schema->getTableSchema('{{testCascadeDropTable2}}') !== null) {
+            $command->dropTable('{{testCascadeDropTable2}}')->execute();
+        }
+        if ($schema->getTableSchema('{{testCascadeDropTable}}') !== null) {
+            $command->dropTable('{{testCascadeDropTable}}')->execute();
+        }
+
+        $command->createTable(
+            '{{testCascadeDropTable}}',
+            ['id' => 'integer not null unique', 'foo' => 'integer'],
+        )->execute();
+        $this->assertNotNull($schema->getTableSchema('{{testCascadeDropTable}}', true));
+
+        $command->createTable(
+            '{{testCascadeDropTable2}}',
+            ['id' => 'integer not null unique', 'foreign_id' => 'integer'],
+        )->execute();
+        $this->assertNotNull($schema->getTableSchema('{{testCascadeDropTable2}}', true));
+
+        $command
+            ->addForeignKey('{{testCascadeDropTable2}}', 'fgk', 'foreign_id', '{{testCascadeDropTable}}', 'id')
+            ->execute();
+        $this->assertNotEmpty($schema->getTableForeignKeys('{{testCascadeDropTable2}}', true));
+
+        $command->dropTable('{{testCascadeDropTable}}', cascade: true)->execute();
+        $this->assertNull($schema->getTableSchema('{{testCascadeDropTable}}', true));
+        $this->assertEmpty($schema->getTableForeignKeys('{{testCascadeDropTable2}}', true));
 
         $db->close();
     }
@@ -2104,5 +2167,50 @@ abstract class CommonCommandTest extends AbstractCommandTest
         $this->assertSame($uuidSource, $uuidString);
 
         $db->close();
+    }
+
+    public function testJsonTable(): void
+    {
+        $db = $this->getConnection();
+        $command = $db->createCommand();
+
+        if ($db->getTableSchema('json_table', true) !== null) {
+            $command->dropTable('json_table')->execute();
+        }
+
+        $command->createTable('json_table', [
+            'id' => PseudoType::PK,
+            'json_col' => ColumnBuilder::json(),
+        ])->execute();
+
+        $command->insert('json_table', ['json_col' => ['a' => 1, 'b' => 2]]);
+
+        $typeHint = $db->getDriverName() === 'pgsql' ? '::jsonb' : '';
+        $expectedValue = $db->getQuoter()->quoteValue('{"a":1,"b":2}') . $typeHint;
+
+        $this->assertSame(
+            DbHelper::replaceQuotes(
+                "INSERT INTO [[json_table]] ([[json_col]]) VALUES (:qp0$typeHint)",
+                $db->getDriverName(),
+            ),
+            $command->getSql()
+        );
+        $this->assertEquals([':qp0' => new Param('{"a":1,"b":2}', DataType::STRING)], $command->getParams(false));
+        $this->assertSame(
+            DbHelper::replaceQuotes(
+                "INSERT INTO [[json_table]] ([[json_col]]) VALUES ($expectedValue)",
+                $db->getDriverName(),
+            ),
+            $command->getRawSql()
+        );
+        $this->assertSame(1, $command->execute());
+
+        $tableSchema = $db->getTableSchema('json_table', true);
+        $this->assertNotNull($tableSchema);
+        $this->assertSame('json_col', $tableSchema->getColumn('json_col')->getName());
+        $this->assertSame(ColumnType::JSON, $tableSchema->getColumn('json_col')->getType());
+
+        $value = (new Query($db))->select('json_col')->from('json_table')->where(['id' => 1])->scalar();
+        $this->assertSame('{"a":1,"b":2}', str_replace(' ', '', $value));
     }
 }